Check if dist directory exists before removing
Apparently I never ran this without the `dist` dir existing, so on a
clean `git clone` I ran into this issue.

Check if the dist directory exists before trying to remove it.

[skip ci]

+++ Rakefile
@@ -338,6 +338,8 @@ end
 task :release => ["release:all"]
 
 def clean_dist_dir
+  return unless Dir.exist? DIST_DIR
+
   FileUtils.remove_dir DIST_DIR
 end
